refactor: move capability workflows out of router
