+++ src/app/routers/operational.py
@@ -1,5 +1,4 @@
 from collections.abc import Sequence
-from dataclasses import dataclass
 
 from fastapi import APIRouter, Request, Response, status
 from prometheus_client import CONTENT_TYPE_LATEST, generate_latest
@@ -8,10 +7,7 @@ from app.api_errors import STANDARD_ERROR_RESPONSES
 from app.contracts.capabilities import (
     CAPABILITY_FEATURE_KEYS,
     CapabilityFeature,
-    CapabilityWorkflow,
     IntegrationCapabilitiesResponse,
-    SupportedInputMode,
-    WorkflowSupportStatus,
 )
 from app.contracts.ops import (
     DependencyStatus,
@@ -29,122 +25,13 @@ from app.service_metadata import (
     SERVICE_VERSION,
     SUPPORTED_INPUT_MODES,
 )
+from app.services.capability_workflows import build_capability_workflows
 from app.trust_telemetry import (
     DeclaredProductTrustTelemetrySnapshot,
     build_declared_product_trust_telemetry_snapshot,
 )
 
 router = APIRouter()
-
-
-@dataclass(frozen=True)
-class _CapabilityWorkflowSpec:
-    workflow_key: str
-    endpoint_path: str
-    supported_input_modes: list[SupportedInputMode]
-    support_status: WorkflowSupportStatus
-    notes: list[str]
-
-
-_CAPABILITY_WORKFLOW_SPECS: tuple[_CapabilityWorkflowSpec, ...] = (
-    _CapabilityWorkflowSpec(
-        workflow_key="risk_snapshot",
-        endpoint_path="/analytics/risk/calculate",
-        supported_input_modes=["stateless", "stateful"],
-        support_status="full",
-        notes=[
-            "simulation is intentionally unsupported",
-            "benchmark-dependent metrics require benchmark returns",
-            "VaR and expected shortfall are signed return-threshold metrics",
-        ],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="concentration_risk",
-        endpoint_path="/analytics/risk/concentration",
-        supported_input_modes=["stateless", "stateful", "simulation"],
-        support_status="full",
-        notes=[
-            "simulation is supported only for concentration risk",
-            "issuer concentration includes coverage diagnostics",
-        ],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="drawdown_analytics",
-        endpoint_path="/analytics/risk/drawdown",
-        supported_input_modes=["stateless", "stateful"],
-        support_status="full",
-        notes=["simulation is intentionally unsupported"],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="rolling_risk_analytics",
-        endpoint_path="/analytics/risk/rolling-metrics",
-        supported_input_modes=["stateless", "stateful"],
-        support_status="full",
-        notes=[
-            "simulation is intentionally unsupported",
-            "stateful rolling Sharpe depends on risk-free series availability from lotus-core",
-        ],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="historical_risk_attribution",
-        endpoint_path="/analytics/risk/historical-attribution",
-        supported_input_modes=["stateless", "stateful"],
-        support_status="partial",
-        notes=[
-            "simulation is intentionally unsupported",
-            "stateful active-risk supports POSITION, SECTOR, ASSET_CLASS, and ISSUER",
-            "issuer active-risk consumes lotus-performance benchmark exposure context issuer groups",
-            "historical-attribution response metadata is the authoritative active-risk support contract",
-            "attribution residual and reconciled_sum must be preserved with contributors",
-        ],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="mandate_risk_health_context",
-        endpoint_path="/analytics/risk/mandate-health-context",
-        supported_input_modes=["stateless"],
-        support_status="partial",
-        notes=[
-            "derives bounded mandate risk health from source-owned tracking-error methodology",
-            "returns threshold posture, lineage, and non-claim reason codes for Manage consumption",
-            "does not create mandate actions, rebalance waves, or client communication",
-        ],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="regime_scenario_pack_evaluation",
-        endpoint_path="/analytics/risk/regime-scenario-pack/evaluate",
-        supported_input_modes=["stateless"],
-        support_status="full",
-        notes=[
-            "evaluates caller-supplied exposure weights against governed CIO scenario packs",
-            "returns source-owned worst-case loss, per-security contribution rows when supplied, CIO approval/effective-period/applicability posture, policy breach posture, and lineage",
-            "does not forecast market states or accept browser-owned scenario methodology",
-        ],
-    ),
-    _CapabilityWorkflowSpec(
-        workflow_key="risk_event_affected_cohort",
-        endpoint_path="/analytics/risk/risk-event-cohorts/evaluate",
-        supported_input_modes=["stateless"],
-        support_status="partial",
-        notes=[
-            "evaluates candidate portfolios against governed risk-event definitions",
-            "returns affected membership, exclusions, impact scores, source refs, and supportability",
-            "does not create rebalance waves or own campaign approval workflow",
-        ],
-    ),
-)
-
-
-def _capability_workflows() -> list[CapabilityWorkflow]:
-    return [
-        CapabilityWorkflow(
-            workflow_key=spec.workflow_key,
-            endpoint_path=spec.endpoint_path,
-            supported_input_modes=list(spec.supported_input_modes),
-            support_status=spec.support_status,
-            notes=list(spec.notes),
-        )
-        for spec in _CAPABILITY_WORKFLOW_SPECS
-    ]
 
 
 def _dependency_statuses(dependencies: Sequence[DependencyRuntimeView]) -> list[DependencyStatus]:
@@ -284,7 +171,7 @@ async def integration_capabilities() -> IntegrationCapabilitiesResponse:
         policy_version="risk.v1",
         supported_input_modes=list(SUPPORTED_INPUT_MODES),
         features=[CapabilityFeature(key=feature_key) for feature_key in CAPABILITY_FEATURE_KEYS],
-        workflows=_capability_workflows(),
+        workflows=build_capability_workflows(),
     )
